unified all methods under a client, which contains the low level sdk and the high-level uploader. Splitted functionality in files and docuemented necessary IAM policy access
