+++ healthcheck.go
@@ -1,11 +1,20 @@
-package s3client
+// file: healthcheck.go
+//
+// Contains methods to get the health state of an S3 client from S3,
+// by checking that the bucket exists in the provided region.
+//
+// Requires "s3:ListBucket" action allowed by IAM policy for the bucket,
+// as defined by `check-{bucketName}-bucket` policies in dp-setup
+package s3
 
 import (
 	"context"
 	"fmt"
 
 	health "github.com/ONSdigital/dp-healthcheck/healthcheck"
+	"github.com/aws/aws-sdk-go/aws"
 	"github.com/aws/aws-sdk-go/aws/awserr"
+	"github.com/aws/aws-sdk-go/service/s3"
 )
 
 // ServiceName S3
@@ -18,7 +27,7 @@ const MsgHealthy = "S3 is healthy"
 
 // Checker validates that the S3 bucket exists, and updates the provided CheckState accordingly.
 // Any error during the state update will be returned
-func (cli *S3) Checker(ctx context.Context, state *health.CheckState) error {
+func (cli *Client) Checker(ctx context.Context, state *health.CheckState) error {
 	if err := cli.ValidateBucket(); err != nil {
 		if awsErr, ok := err.(awserr.Error); ok {
 			// AWS error
@@ -34,7 +43,7 @@ func (cli *S3) Checker(ctx context.Context, state *health.CheckState) error {
 // handleAWSErr updates the provided CheckState with a Critical state and a message according to the provided AWS error.
 // For inexistent buckets, a relevant error message will be generated, for any other error we use the AWS Code (consice string).
 // Any error during the state update will be returned
-func (cli *S3) handleAWSErr(err awserr.Error, state *health.CheckState) error {
+func (cli *Client) handleAWSErr(err awserr.Error, state *health.CheckState) error {
 	code := err.Code()
 	switch code {
 	case codeNotFound:
@@ -44,4 +53,15 @@ func (cli *S3) handleAWSErr(err awserr.Error, state *health.CheckState) error {
 		// Other AWS error
 		return state.Update(health.StatusCritical, err.Code(), 0)
 	}
+}
+
+// ValidateBucket checks that the bucket exists and returns an error if it
+// does not exist or there was some other error trying to get this information.
+func (cli *Client) ValidateBucket() error {
+	_, err := cli.sdkClient.HeadBucket(
+		&s3.HeadBucketInput{
+			Bucket: aws.String(cli.bucketName),
+		},
+	)
+	return err
 }
