+++ healthcheck_test.go
@@ -1,4 +1,4 @@
-package s3client_test
+package s3_test
 
 import (
 	"context"
@@ -6,7 +6,7 @@ import (
 	"testing"
 
 	health "github.com/ONSdigital/dp-healthcheck/healthcheck"
-	s3client "github.com/ONSdigital/dp-s3"
+	dps3 "github.com/ONSdigital/dp-s3"
 	"github.com/ONSdigital/dp-s3/mock"
 	"github.com/aws/aws-sdk-go/service/s3"
 	. "github.com/smartystreets/goconvey/convey"
@@ -52,20 +52,20 @@ func bucketInexistentRegion(*s3.HeadBucketInput) (*s3.HeadBucketOutput, error) {
 func TestBucketOk(t *testing.T) {
 	Convey("Given that S3 client is available, bucket exists and it was created in the same region as the S3 client config", t, func() {
 
-		// Create S3Client with SDK Mock for existing bucket
+		// Create S3 client with SDK Mock for existing bucket
 		sdkMock := &mock.S3SDKClientMock{
 			HeadBucketFunc: bucketExists,
 		}
-		s3Cli := s3client.InstantiateClient(sdkMock, nil, ExistingBucket, ExpectedRegion, nil)
+		cli := dps3.InstantiateClient(sdkMock, nil, nil, nil, ExistingBucket, ExpectedRegion, nil)
 
 		// CheckState for test validation
-		checkState := health.NewCheckState(s3client.ServiceName)
+		checkState := health.NewCheckState(dps3.ServiceName)
 
 		Convey("Checker updates the CheckState to a successful state", func() {
-			s3Cli.Checker(context.Background(), checkState)
+			cli.Checker(context.Background(), checkState)
 			So(len(sdkMock.HeadBucketCalls()), ShouldEqual, 1)
 			So(checkState.Status(), ShouldEqual, health.StatusOK)
-			So(checkState.Message(), ShouldEqual, s3client.MsgHealthy)
+			So(checkState.Message(), ShouldEqual, dps3.MsgHealthy)
 			So(checkState.StatusCode(), ShouldEqual, 0)
 		})
 	})
@@ -74,17 +74,17 @@ func TestBucketOk(t *testing.T) {
 func TestBucketDoesNotExist(t *testing.T) {
 	Convey("Given that S3 client is available and bucket does not exist", t, func() {
 
-		// Create S3Client with SDK Mock for inexistent bucket
+		// Create S3 client with SDK Mock for inexistent bucket
 		sdkMock := &mock.S3SDKClientMock{
 			HeadBucketFunc: bucketDoesNotExist,
 		}
-		s3Cli := s3client.InstantiateClient(sdkMock, nil, InexistentBucket, ExpectedRegion, nil)
+		cli := dps3.InstantiateClient(sdkMock, nil, nil, nil, InexistentBucket, ExpectedRegion, nil)
 
 		// CheckState for test validation
-		checkState := health.NewCheckState(s3client.ServiceName)
+		checkState := health.NewCheckState(dps3.ServiceName)
 
 		Convey("Checker updates the CheckState to a critical state with the relevant error message", func() {
-			s3Cli.Checker(context.Background(), checkState)
+			cli.Checker(context.Background(), checkState)
 			So(len(sdkMock.HeadBucketCalls()), ShouldEqual, 1)
 			So(checkState.Status(), ShouldEqual, health.StatusCritical)
 			So(checkState.Message(), ShouldEqual, msgBucketNotFound)
@@ -96,17 +96,17 @@ func TestBucketDoesNotExist(t *testing.T) {
 func TestBucketUnexpectedRegion(t *testing.T) {
 	Convey("Given that S3 client is available and bucket was created in a different region than the S3 client config", t, func() {
 
-		// Create S3Client with SDK Mock for unexpected region for bucket
+		// Create S3 client with SDK Mock for unexpected region for bucket
 		sdkMock := &mock.S3SDKClientMock{
 			HeadBucketFunc: bucketWrongRegion,
 		}
-		s3Cli := s3client.InstantiateClient(sdkMock, nil, ExistingBucket, UnexpectedRegion, nil)
+		cli := dps3.InstantiateClient(sdkMock, nil, nil, nil, ExistingBucket, UnexpectedRegion, nil)
 
 		// CheckState for test validation
-		checkState := health.NewCheckState(s3client.ServiceName)
+		checkState := health.NewCheckState(dps3.ServiceName)
 
 		Convey("Checker updates the CheckState to a critical state with the relevant error message", func() {
-			s3Cli.Checker(context.Background(), checkState)
+			cli.Checker(context.Background(), checkState)
 			So(len(sdkMock.HeadBucketCalls()), ShouldEqual, 1)
 			So(checkState.Status(), ShouldEqual, health.StatusCritical)
 			So(checkState.Message(), ShouldEqual, msgWrongRegion)
@@ -118,17 +118,17 @@ func TestBucketUnexpectedRegion(t *testing.T) {
 func TestBucketInexistentRegion(t *testing.T) {
 	Convey("Given that S3 client is available, bucket exists, but S3 is configured with an inexistent region", t, func() {
 
-		// Create S3Client with SDK Mock for inexistent region
+		// Create S3 client with SDK Mock for inexistent region
 		sdkMock := &mock.S3SDKClientMock{
 			HeadBucketFunc: bucketInexistentRegion,
 		}
-		s3Cli := s3client.InstantiateClient(sdkMock, nil, ExistingBucket, InexistentRegion, nil)
+		cli := dps3.InstantiateClient(sdkMock, nil, nil, nil, ExistingBucket, InexistentRegion, nil)
 
 		// CheckState for test validation
-		checkState := health.NewCheckState(s3client.ServiceName)
+		checkState := health.NewCheckState(dps3.ServiceName)
 
 		Convey("Checker updates the CheckState to a critical state with the relevant error message", func() {
-			s3Cli.Checker(context.Background(), checkState)
+			cli.Checker(context.Background(), checkState)
 			So(len(sdkMock.HeadBucketCalls()), ShouldEqual, 1)
 			expectedErr := errors.New(msgInexistentRegion)
 			So(checkState.Status(), ShouldEqual, health.StatusCritical)
