+++ upload_test.go
@@ -1,4 +1,4 @@
-package s3client_test
+package s3_test
 
 import (
 	"context"
@@ -6,7 +6,7 @@ import (
 	"fmt"
 	"testing"
 
-	s3client "github.com/ONSdigital/dp-s3"
+	dps3 "github.com/ONSdigital/dp-s3"
 	"github.com/ONSdigital/dp-s3/mock"
 	"github.com/ONSdigital/log.go/v2/log"
 	"github.com/aws/aws-sdk-go/service/s3/s3manager"
@@ -19,26 +19,24 @@ var (
 )
 
 func TestUpload(t *testing.T) {
-	Convey("Given an Uploader configured with an sdk uploader", t, func() {
-		sdkMock := &mock.S3SDKClientMock{}
-		s3Cli := s3client.InstantiateClient(sdkMock, nil, testBucket, ExpectedRegion, nil)
+	Convey("Given a client configured with a successful SDK uploader", t, func() {
 		sdkUploaderMock := &mock.S3SDKUploaderMock{
 			UploadFunc: func(in *s3manager.UploadInput, options ...func(*s3manager.Uploader)) (*s3manager.UploadOutput, error) {
 				return nil, nil
 			},
 		}
-		uploader := s3client.InstantiateUploader(s3Cli, sdkUploaderMock, nil)
+		cli := dps3.InstantiateClient(nil, nil, sdkUploaderMock, nil, testBucket, ExpectedRegion, nil)
 
 		Convey("Calling Upload with a valid s3 key results in sdk Upload being called as expected", func() {
-			_, err := uploader.Upload(&s3manager.UploadInput{Key: &testS3Key})
+			_, err := cli.Upload(&s3manager.UploadInput{Key: &testS3Key})
 			So(err, ShouldBeNil)
 			So(len(sdkUploaderMock.UploadCalls()), ShouldEqual, 1)
 			So(*sdkUploaderMock.UploadCalls()[0].In.Bucket, ShouldEqual, testBucket)
 		})
 
 		Convey("Calling Upload with nil input returns the expected error", func() {
-			_, err := uploader.Upload(nil)
-			So(err, ShouldResemble, s3client.NewError(
+			_, err := cli.Upload(nil)
+			So(err, ShouldResemble, dps3.NewError(
 				fmt.Errorf("validation error for Upload: %w",
 					errors.New("nil input provided"),
 				),
@@ -50,20 +48,18 @@ func TestUpload(t *testing.T) {
 		})
 	})
 
-	Convey("Given an SDK Uploader that fails to upload", t, func() {
+	Convey("Given a client configured with an SDK uploader that fails to upload", t, func() {
 		errUploader := errors.New("failed to upload file")
-		sdkMock := &mock.S3SDKClientMock{}
-		s3Cli := s3client.InstantiateClient(sdkMock, nil, testBucket, ExpectedRegion, nil)
 		sdkUploaderMock := &mock.S3SDKUploaderMock{
 			UploadFunc: func(in *s3manager.UploadInput, options ...func(*s3manager.Uploader)) (*s3manager.UploadOutput, error) {
 				return nil, errUploader
 			},
 		}
-		uploader := s3client.InstantiateUploader(s3Cli, sdkUploaderMock, nil)
+		cli := dps3.InstantiateClient(nil, nil, sdkUploaderMock, nil, testBucket, ExpectedRegion, nil)
 
 		Convey("Calling Upload with a valid s3 key results in the expected error being returned", func() {
-			_, err := uploader.Upload(&s3manager.UploadInput{Key: &testS3Key})
-			So(err, ShouldResemble, s3client.NewError(
+			_, err := cli.Upload(&s3manager.UploadInput{Key: &testS3Key})
+			So(err, ShouldResemble, dps3.NewError(
 				fmt.Errorf("failed to upload: %w", errUploader),
 				log.Data{
 					"bucket_name": testBucket,
@@ -79,18 +75,16 @@ func TestUpload(t *testing.T) {
 func TestUploadWithContext(t *testing.T) {
 	ctx := context.Background()
 
-	Convey("Given an Uploader configured with an sdk uploader", t, func() {
-		sdkMock := &mock.S3SDKClientMock{}
-		s3Cli := s3client.InstantiateClient(sdkMock, nil, testBucket, ExpectedRegion, nil)
+	Convey("Given a client configured with a successful SDK uploader", t, func() {
 		sdkUploaderMock := &mock.S3SDKUploaderMock{
 			UploadWithContextFunc: func(ctx context.Context, in *s3manager.UploadInput, options ...func(*s3manager.Uploader)) (*s3manager.UploadOutput, error) {
 				return nil, nil
 			},
 		}
-		uploader := s3client.InstantiateUploader(s3Cli, sdkUploaderMock, nil)
+		cli := dps3.InstantiateClient(nil, nil, sdkUploaderMock, nil, testBucket, ExpectedRegion, nil)
 
 		Convey("Calling UploadWithContext with a valid s3 key and context results in sdk UploadWithContext being called as expected", func() {
-			_, err := uploader.UploadWithContext(ctx, &s3manager.UploadInput{Key: &testS3Key})
+			_, err := cli.UploadWithContext(ctx, &s3manager.UploadInput{Key: &testS3Key})
 			So(err, ShouldBeNil)
 			So(len(sdkUploaderMock.UploadWithContextCalls()), ShouldEqual, 1)
 			So(*sdkUploaderMock.UploadWithContextCalls()[0].In.Bucket, ShouldEqual, testBucket)
@@ -98,8 +92,8 @@ func TestUploadWithContext(t *testing.T) {
 		})
 
 		Convey("Calling UploadWithContext with nil context returns the expected error", func() {
-			_, err := uploader.UploadWithContext(nil, &s3manager.UploadInput{Key: &testS3Key})
-			So(err, ShouldResemble, s3client.NewError(
+			_, err := cli.UploadWithContext(nil, &s3manager.UploadInput{Key: &testS3Key})
+			So(err, ShouldResemble, dps3.NewError(
 				errors.New("nil context provided to UploadWithContext"),
 				log.Data{
 					"bucket_name": testBucket,
@@ -110,8 +104,8 @@ func TestUploadWithContext(t *testing.T) {
 		})
 
 		Convey("Calling UploadWithContext with nil input returns the expected error", func() {
-			_, err := uploader.UploadWithContext(ctx, nil)
-			So(err, ShouldResemble, s3client.NewError(
+			_, err := cli.UploadWithContext(ctx, nil)
+			So(err, ShouldResemble, dps3.NewError(
 				fmt.Errorf("validation error for UploadWithContext: %w",
 					errors.New("nil input provided"),
 				),
@@ -123,20 +117,18 @@ func TestUploadWithContext(t *testing.T) {
 		})
 	})
 
-	Convey("Given an SDK Uploader that fails to upload", t, func() {
+	Convey("Given a client configured with an SDK uploader that fails to upload", t, func() {
 		errUploader := errors.New("failed to upload file")
-		sdkMock := &mock.S3SDKClientMock{}
-		s3Cli := s3client.InstantiateClient(sdkMock, nil, testBucket, ExpectedRegion, nil)
 		sdkUploaderMock := &mock.S3SDKUploaderMock{
 			UploadWithContextFunc: func(ctx context.Context, in *s3manager.UploadInput, options ...func(*s3manager.Uploader)) (*s3manager.UploadOutput, error) {
 				return nil, errUploader
 			},
 		}
-		uploader := s3client.InstantiateUploader(s3Cli, sdkUploaderMock, nil)
+		cli := dps3.InstantiateClient(nil, nil, sdkUploaderMock, nil, testBucket, ExpectedRegion, nil)
 
 		Convey("Calling UploadWithContext with a valid s3 key and context results in the expected error being returned", func() {
-			_, err := uploader.UploadWithContext(ctx, &s3manager.UploadInput{Key: &testS3Key})
-			So(err, ShouldResemble, s3client.NewError(
+			_, err := cli.UploadWithContext(ctx, &s3manager.UploadInput{Key: &testS3Key})
+			So(err, ShouldResemble, dps3.NewError(
 				fmt.Errorf("failed to upload with context: %w", errUploader),
 				log.Data{
 					"bucket_name": testBucket,
@@ -151,19 +143,17 @@ func TestUploadWithContext(t *testing.T) {
 }
 
 func TestUploadWithPSK(t *testing.T) {
-	Convey("Given an Uploader configured with user-defined psk", t, func() {
+	Convey("Given a client configured with a user-defined psk uploader", t, func() {
 		psk := []byte("test psk")
-		sdkMock := &mock.S3SDKClientMock{}
-		s3Cli := s3client.InstantiateClient(sdkMock, nil, testBucket, ExpectedRegion, nil)
 		cryptoUploaderMock := &mock.S3CryptoUploaderMock{
 			UploadWithPSKFunc: func(ctx context.Context, in *s3manager.UploadInput, psk []byte) (*s3manager.UploadOutput, error) {
 				return &s3manager.UploadOutput{}, nil
 			},
 		}
-		uploader := s3client.InstantiateUploader(s3Cli, nil, cryptoUploaderMock)
+		cli := dps3.InstantiateClient(nil, nil, nil, cryptoUploaderMock, testBucket, ExpectedRegion, nil)
 
 		Convey("Calling UploadWithPSK with a valid s3 key results in crypto UploadWithPSK being called as expected", func() {
-			_, err := uploader.UploadWithPSK(&s3manager.UploadInput{Key: &testS3Key}, psk)
+			_, err := cli.UploadWithPSK(&s3manager.UploadInput{Key: &testS3Key}, psk)
 			So(err, ShouldBeNil)
 			So(len(cryptoUploaderMock.UploadWithPSKCalls()), ShouldEqual, 1)
 			So(*cryptoUploaderMock.UploadWithPSKCalls()[0].In.Bucket, ShouldEqual, testBucket)
@@ -171,8 +161,8 @@ func TestUploadWithPSK(t *testing.T) {
 		})
 
 		Convey("Calling UploadWithPSK with nil psk returns the expected error", func() {
-			_, err := uploader.UploadWithPSK(&s3manager.UploadInput{Key: &testS3Key}, nil)
-			So(err, ShouldResemble, s3client.NewError(
+			_, err := cli.UploadWithPSK(&s3manager.UploadInput{Key: &testS3Key}, nil)
+			So(err, ShouldResemble, dps3.NewError(
 				errors.New("nil or empty psk provided to UploadWithPSK"),
 				log.Data{
 					"bucket_name": testBucket,
@@ -183,8 +173,8 @@ func TestUploadWithPSK(t *testing.T) {
 		})
 
 		Convey("Calling UploadWithPSK with nil input returns the expected error", func() {
-			_, err := uploader.UploadWithPSK(nil, psk)
-			So(err, ShouldResemble, s3client.NewError(
+			_, err := cli.UploadWithPSK(nil, psk)
+			So(err, ShouldResemble, dps3.NewError(
 				fmt.Errorf("validation error for UploadWithPSK: %w",
 					errors.New("nil input provided"),
 				),
@@ -196,21 +186,19 @@ func TestUploadWithPSK(t *testing.T) {
 		})
 	})
 
-	Convey("Given an Uploader configured with user-defined psk that fails to upload", t, func() {
+	Convey("Given a client configured with a user-defined psk uploader that fails to upload", t, func() {
 		errCryptoUploader := errors.New("failed to upload file")
 		psk := []byte("test psk")
-		sdkMock := &mock.S3SDKClientMock{}
-		s3Cli := s3client.InstantiateClient(sdkMock, nil, testBucket, ExpectedRegion, nil)
 		cryptoUploaderMock := &mock.S3CryptoUploaderMock{
 			UploadWithPSKFunc: func(ctx context.Context, in *s3manager.UploadInput, psk []byte) (*s3manager.UploadOutput, error) {
 				return nil, errCryptoUploader
 			},
 		}
-		uploader := s3client.InstantiateUploader(s3Cli, nil, cryptoUploaderMock)
+		cli := dps3.InstantiateClient(nil, nil, nil, cryptoUploaderMock, testBucket, ExpectedRegion, nil)
 
 		Convey("Calling UploadWithPSK with a valid s3 key and context results in the expected error being returned", func() {
-			_, err := uploader.UploadWithPSK(&s3manager.UploadInput{Key: &testS3Key}, psk)
-			So(err, ShouldResemble, s3client.NewError(
+			_, err := cli.UploadWithPSK(&s3manager.UploadInput{Key: &testS3Key}, psk)
+			So(err, ShouldResemble, dps3.NewError(
 				fmt.Errorf("failed to upload with psk: %w", errCryptoUploader),
 				log.Data{
 					"bucket_name": testBucket,
@@ -227,19 +215,17 @@ func TestUploadWithPSK(t *testing.T) {
 func TestUploadWithPSKAndContext(t *testing.T) {
 	ctx := context.Background()
 
-	Convey("Given an Uploader configured with user-defined psk", t, func() {
+	Convey("Given a client configured with a user-defined psk uploader", t, func() {
 		psk := []byte("test psk")
-		sdkMock := &mock.S3SDKClientMock{}
-		s3Cli := s3client.InstantiateClient(sdkMock, nil, testBucket, ExpectedRegion, nil)
 		cryptoUploaderMock := &mock.S3CryptoUploaderMock{
 			UploadWithPSKFunc: func(ctx context.Context, in *s3manager.UploadInput, psk []byte) (*s3manager.UploadOutput, error) {
 				return &s3manager.UploadOutput{}, nil
 			},
 		}
-		uploader := s3client.InstantiateUploader(s3Cli, nil, cryptoUploaderMock)
+		cli := dps3.InstantiateClient(nil, nil, nil, cryptoUploaderMock, testBucket, ExpectedRegion, nil)
 
 		Convey("Calling UploadWithPSKAndContext with a valid s3 key and context results in crypto UploadWithPSK being called as expected", func() {
-			_, err := uploader.UploadWithPSKAndContext(ctx, &s3manager.UploadInput{Key: &testS3Key}, psk)
+			_, err := cli.UploadWithPSKAndContext(ctx, &s3manager.UploadInput{Key: &testS3Key}, psk)
 			So(err, ShouldBeNil)
 			So(len(cryptoUploaderMock.UploadWithPSKCalls()), ShouldEqual, 1)
 			So(*cryptoUploaderMock.UploadWithPSKCalls()[0].In.Bucket, ShouldEqual, testBucket)
@@ -247,8 +233,8 @@ func TestUploadWithPSKAndContext(t *testing.T) {
 		})
 
 		Convey("Calling UploadWithPSKAndContext with nil context returns the expected error", func() {
-			_, err := uploader.UploadWithPSKAndContext(nil, &s3manager.UploadInput{Key: &testS3Key}, psk)
-			So(err, ShouldResemble, s3client.NewError(
+			_, err := cli.UploadWithPSKAndContext(nil, &s3manager.UploadInput{Key: &testS3Key}, psk)
+			So(err, ShouldResemble, dps3.NewError(
 				errors.New("nil context provided to UploadWithPSKAndContext"),
 				log.Data{
 					"bucket_name": testBucket,
@@ -259,8 +245,8 @@ func TestUploadWithPSKAndContext(t *testing.T) {
 		})
 
 		Convey("Calling UploadWithPSKAndContext with nil psk returns the expected error", func() {
-			_, err := uploader.UploadWithPSKAndContext(ctx, &s3manager.UploadInput{Key: &testS3Key}, nil)
-			So(err, ShouldResemble, s3client.NewError(
+			_, err := cli.UploadWithPSKAndContext(ctx, &s3manager.UploadInput{Key: &testS3Key}, nil)
+			So(err, ShouldResemble, dps3.NewError(
 				errors.New("nil or empty psk provided to UploadWithPSKAndContext"),
 				log.Data{
 					"bucket_name": testBucket,
@@ -271,8 +257,8 @@ func TestUploadWithPSKAndContext(t *testing.T) {
 		})
 
 		Convey("Calling UploadWithPSKAndContext with nil input returns the expected error", func() {
-			_, err := uploader.UploadWithPSKAndContext(ctx, nil, psk)
-			So(err, ShouldResemble, s3client.NewError(
+			_, err := cli.UploadWithPSKAndContext(ctx, nil, psk)
+			So(err, ShouldResemble, dps3.NewError(
 				fmt.Errorf("validation error for UploadWithPSKAndContext: %w",
 					errors.New("nil input provided"),
 				),
@@ -284,21 +270,19 @@ func TestUploadWithPSKAndContext(t *testing.T) {
 		})
 	})
 
-	Convey("Given an Uploader configured with user-defined psk that fails to upload", t, func() {
+	Convey("Given a client configured with a user-defined psk uploader that fails to upload", t, func() {
 		errCryptoUploader := errors.New("failed to upload file")
 		psk := []byte("test psk")
-		sdkMock := &mock.S3SDKClientMock{}
-		s3Cli := s3client.InstantiateClient(sdkMock, nil, testBucket, ExpectedRegion, nil)
 		cryptoUploaderMock := &mock.S3CryptoUploaderMock{
 			UploadWithPSKFunc: func(ctx context.Context, in *s3manager.UploadInput, psk []byte) (*s3manager.UploadOutput, error) {
 				return nil, errCryptoUploader
 			},
 		}
-		uploader := s3client.InstantiateUploader(s3Cli, nil, cryptoUploaderMock)
+		cli := dps3.InstantiateClient(nil, nil, nil, cryptoUploaderMock, testBucket, ExpectedRegion, nil)
 
 		Convey("Calling UploadWithPSKAndContext with a valid s3 key and context results in the expected error being returned", func() {
-			_, err := uploader.UploadWithPSKAndContext(ctx, &s3manager.UploadInput{Key: &testS3Key}, psk)
-			So(err, ShouldResemble, s3client.NewError(
+			_, err := cli.UploadWithPSKAndContext(ctx, &s3manager.UploadInput{Key: &testS3Key}, psk)
+			So(err, ShouldResemble, dps3.NewError(
 				fmt.Errorf("failed to upload with psk: %w", errCryptoUploader),
 				log.Data{
 					"bucket_name": testBucket,
@@ -313,12 +297,10 @@ func TestUploadWithPSKAndContext(t *testing.T) {
 }
 
 func TestValidateInput(t *testing.T) {
-	u := s3client.Uploader{
-		S3: s3client.InstantiateClient(nil, nil, testBucket, "", nil),
-	}
+	cli := dps3.InstantiateClient(nil, nil, nil, nil, testBucket, "", nil)
 
 	Convey("validating an input with only an s3 key is successful with the expected logData", t, func() {
-		logData, err := u.ValidateInput(&s3manager.UploadInput{
+		logData, err := cli.ValidateUploadInput(&s3manager.UploadInput{
 			Key: &testS3Key,
 		})
 
@@ -330,7 +312,7 @@ func TestValidateInput(t *testing.T) {
 	})
 
 	Convey("validating an input with an s3 key and the expected bucket is successful with the expected logData", t, func() {
-		logData, err := u.ValidateInput(&s3manager.UploadInput{
+		logData, err := cli.ValidateUploadInput(&s3manager.UploadInput{
 			Key:    &testS3Key,
 			Bucket: &testBucket,
 		})
@@ -343,7 +325,7 @@ func TestValidateInput(t *testing.T) {
 	})
 
 	Convey("validating a nil input fails with the expected error and logData", t, func() {
-		logData, err := u.ValidateInput(nil)
+		logData, err := cli.ValidateUploadInput(nil)
 
 		So(err, ShouldResemble, errors.New("nil input provided"))
 		So(logData, ShouldResemble, log.Data{
@@ -352,7 +334,7 @@ func TestValidateInput(t *testing.T) {
 	})
 
 	Convey("validating an input without s3 key fails with the expected error and logData", t, func() {
-		logData, err := u.ValidateInput(&s3manager.UploadInput{})
+		logData, err := cli.ValidateUploadInput(&s3manager.UploadInput{})
 
 		So(err, ShouldResemble, errors.New("nil or empty s3 key provided in input"))
 		So(logData, ShouldResemble, log.Data{
@@ -362,7 +344,7 @@ func TestValidateInput(t *testing.T) {
 
 	Convey("validating an input without s3 key fails with the expected error and logData", t, func() {
 		emptyKey := ""
-		logData, err := u.ValidateInput(&s3manager.UploadInput{
+		logData, err := cli.ValidateUploadInput(&s3manager.UploadInput{
 			Key: &emptyKey,
 		})
 		So(err, ShouldResemble, errors.New("nil or empty s3 key provided in input"))
@@ -373,7 +355,7 @@ func TestValidateInput(t *testing.T) {
 
 	Convey("validating an input with an s3 key and an unexpected bucket fails with the expected error and logData", t, func() {
 		otherBucket := "otherBucket"
-		logData, err := u.ValidateInput(&s3manager.UploadInput{
+		logData, err := cli.ValidateUploadInput(&s3manager.UploadInput{
 			Key:    &testS3Key,
 			Bucket: &otherBucket,
 		})
